requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==4.1.2
django-filter==22.1
django-mysql==4.7.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
packaging==21.3
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.9.0
